+++ src/pages/login/index.tsx
@@ -1,9 +1,8 @@
-'use client';
-
 import { styled } from '@mui/material';
-import SignUpForm from '@/app/(plain)/(guest)/signup/_components/SignUpForm';
+import LoginForm from '@/app/(plain)/(guest)/login/_components/LoginForm';
+import { NextPageWithLayout } from 'next';
 
-const Root = styled('div')`
+const StyledLogin = styled('div')`
   display: grid;
   min-height: 100vh;
   grid-template-columns: 2fr 3fr;
@@ -22,7 +21,7 @@ const Root = styled('div')`
   }
 `;
 
-const Left = styled('div')`
+const StyledLeftColumn = styled('div')`
   display: flex;
   flex-direction: column;
   background-color: ${({ theme }) => theme.palette.primary.main};
@@ -38,14 +37,19 @@ const Left = styled('div')`
   }
 `;
 
-const LeftTitle = styled('h1')`
+const StyledTitle = styled('h1')`
   font-size: 40px;
   margin-block-start: 0;
   margin-block-end: 2%;
   color: ${({ theme }) => theme.palette.primary.contrastText};
 `;
 
-const Right = styled('div')`
+const StyledText = styled('span')`
+  font-size: 16px;
+  color: ${({ theme }) => theme.palette.primary.contrastText};
+`;
+
+const StyledRightColumn = styled('div')`
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -61,18 +65,25 @@ const Title = styled('div')`
   }
 `;
 
-export default function page() {
+const Page: NextPageWithLayout = () => {
   return (
     <main>
-      <Root>
-        <Left>
-          <LeftTitle>アカウントを作成</LeftTitle>
-        </Left>
-        <Title>Cuculus アカウント登録</Title>
-        <Right>
-          <SignUpForm />
-        </Right>
-      </Root>
+      <StyledLogin>
+        <StyledLeftColumn>
+          <StyledTitle>ログイン</StyledTitle>
+          <StyledText>
+            メールアドレス、パスワードを入力してください。
+          </StyledText>
+        </StyledLeftColumn>
+        <Title>Cuculusにログイン</Title>
+        <StyledRightColumn>
+          <LoginForm />
+        </StyledRightColumn>
+      </StyledLogin>
     </main>
   );
-}
+};
+
+Page.accessLevel = 'guest';
+
+export default Page;
