+++ src/pages/signup/index.tsx
@@ -1,9 +1,8 @@
-'use client';
-
 import { styled } from '@mui/material';
-import LoginForm from '@/app/(plain)/(guest)/login/_components/LoginForm';
+import SignUpForm from '@/app/(plain)/(guest)/signup/_components/SignUpForm';
+import { NextPageWithLayout } from 'next';
 
-const StyledLogin = styled('div')`
+const Root = styled('div')`
   display: grid;
   min-height: 100vh;
   grid-template-columns: 2fr 3fr;
@@ -22,7 +21,7 @@ const StyledLogin = styled('div')`
   }
 `;
 
-const StyledLeftColumn = styled('div')`
+const Left = styled('div')`
   display: flex;
   flex-direction: column;
   background-color: ${({ theme }) => theme.palette.primary.main};
@@ -38,19 +37,14 @@ const StyledLeftColumn = styled('div')`
   }
 `;
 
-const StyledTitle = styled('h1')`
+const LeftTitle = styled('h1')`
   font-size: 40px;
   margin-block-start: 0;
   margin-block-end: 2%;
   color: ${({ theme }) => theme.palette.primary.contrastText};
 `;
 
-const StyledText = styled('span')`
-  font-size: 16px;
-  color: ${({ theme }) => theme.palette.primary.contrastText};
-`;
-
-const StyledRightColumn = styled('div')`
+const Right = styled('div')`
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -66,21 +60,22 @@ const Title = styled('div')`
   }
 `;
 
-export default function page() {
+const Page: NextPageWithLayout = () => {
   return (
     <main>
-      <StyledLogin>
-        <StyledLeftColumn>
-          <StyledTitle>ログイン</StyledTitle>
-          <StyledText>
-            メールアドレス、パスワードを入力してください。
-          </StyledText>
-        </StyledLeftColumn>
-        <Title>Cuculusにログイン</Title>
-        <StyledRightColumn>
-          <LoginForm />
-        </StyledRightColumn>
-      </StyledLogin>
+      <Root>
+        <Left>
+          <LeftTitle>アカウントを作成</LeftTitle>
+        </Left>
+        <Title>Cuculus アカウント登録</Title>
+        <Right>
+          <SignUpForm />
+        </Right>
+      </Root>
     </main>
   );
-}
+};
+
+Page.accessLevel = 'guest';
+
+export default Page;
